Exam/requirements.txt
apache-airflow==2.9.1
pandas==2.1.4
scikit-learn==1.4.2
joblib==1.4.2
